Add files via upload
This repository includes the input sounding, the namelist settings, as well as the model source code (onefile.F) used for a CBL simulation that produces dust-devil-like vortices.  Also included are modifications to the parcel code, which allow one to diagnose the different terms of the vorticity equation along the parcel trajectories.